requirements.txt
Flask
gunicorn
qrcode
psycopg2-binary
